add: Redirect  STDIO to a file when run the compiled program

+++ Source/devExec.pas
@@ -25,6 +25,16 @@ uses
   Windows, Classes, Forms, devcfg, utils;
 
 type
+  TPipeInputThread = class(TThread)
+  private
+    procedure PipeInput;
+    procedure WriteEndLine;
+  public
+    WriteHandle : THandle;
+    InputFile: string;
+    procedure Execute; override;
+  end;
+
   TExecThread = class(TThread)
   private
     fFile: AnsiString;
@@ -35,6 +45,10 @@ type
     fVisible: boolean;
     procedure ExecAndWait;
   public
+    InputWrite : THandle;
+    InputRead : THandle;
+    StartupEvent: THandle;
+    RedirectInput : boolean;
     procedure Execute; override;
   published
     property FileName: AnsiString read fFile write fFile;
@@ -48,6 +62,7 @@ type
   TdevExecutor = class(TPersistent)
   private
     fExec: TExecThread;
+    fPipe: TPipeInputThread;
     fIsRunning: boolean;
     fOnTermEvent: TNotifyEvent;
     procedure TerminateEvent(Sender: TObject);
@@ -55,8 +70,8 @@ type
   public
     destructor Destroy; override;
     procedure Reset;
-    procedure ExecuteAndWatch(sFileName, sParams, sPath: AnsiString; bVisible: boolean; iTimeOut: Cardinal; OnTermEvent:
-      TNotifyEvent);
+    procedure ExecuteAndWatch(sFileName, sParams, sPath: AnsiString; bVisible: boolean;
+      bRedirectInput:boolean; InputFile: string; iTimeOut: Cardinal; OnTermEvent: TNotifyEvent);
   published
     property Running: boolean read fIsRunning;
   end;
@@ -66,7 +81,7 @@ function devExecutor: TdevExecutor;
 implementation
 
 uses
-  main;
+  main,sysutils,dialogs;
 
 { TExecThread }
 
@@ -80,6 +95,7 @@ procedure TExecThread.ExecAndWait;
 var
   StartupInfo: TStartupInfo;
   ProcessInfo: TProcessInformation;
+  sa: TSecurityAttributes;
 begin
   FillChar(StartupInfo, SizeOf(TStartupInfo), 0);
   with StartupInfo do begin
@@ -90,12 +106,38 @@ begin
     else
       wShowWindow := SW_HIDE;
   end;
-  if CreateProcess(nil, PAnsiChar('"' + fFile + '" ' + fParams), nil, nil, False, NORMAL_PRIORITY_CLASS, nil,
-    PAnsiChar(fPath),
-    StartupInfo, ProcessInfo) then begin
-    fProcess := ProcessInfo.hProcess;
-    WaitForSingleObject(ProcessInfo.hProcess, fTimeOut);
+  if RedirectInput then begin
+    // Set up the security attributes struct.
+    sa.nLength := sizeof(TSecurityAttributes);
+    sa.lpSecurityDescriptor := nil;
+    sa.bInheritHandle := true;
+
+    // Create the child input pipe.
+    if not CreatePipe(InputRead, InputWrite, @sa, 0) then begin
+      SetEvent(StartupEvent);
+      Exit;
+    end;
+    {
+    if not SetHandleInformation(fInputwrite, HANDLE_FLAG_INHERIT, 0) then
+      Exit;
+    }
+    StartupInfo.dwFlags := StartupInfo.dwFlags or STARTF_USESTDHANDLES;
+    StartupInfo.hStdInput := InputRead;
+    StartupInfo.hStdOutput := GetStdHandle(STD_OUTPUT_HANDLE);
+    StartupInfo.hStdError := GetStdHandle(STD_ERROR_HANDLE);
   end;
+
+  if CreateProcess(nil, PAnsiChar('"' + fFile + '" ' + fParams), nil, nil, True,
+    NORMAL_PRIORITY_CLASS , nil,
+    PAnsiChar(fPath), StartupInfo, ProcessInfo) then begin
+    fProcess := ProcessInfo.hProcess;
+    SetEvent(StartupEvent);
+    WaitForSingleObject(ProcessInfo.hProcess, fTimeOut);
+  end else
+    SetEvent(StartupEvent);
+
+  if RedirectInput then
+    CloseHandle(InputRead);
   CloseHandle(ProcessInfo.hProcess);
   CloseHandle(ProcessInfo.hThread);
 end;
@@ -124,10 +166,12 @@ begin
 end;
 
 procedure TdevExecutor.ExecuteAndWatch(sFileName, sParams, sPath: AnsiString;
-  bVisible: boolean; iTimeOut: Cardinal; OnTermEvent: TNotifyEvent);
+  bVisible: boolean; bRedirectInput:boolean; InputFile: string;
+  iTimeOut: Cardinal; OnTermEvent: TNotifyEvent);
 begin
   fIsRunning := True;
   fOnTermEvent := OnTermEvent;
+
   fExec := TExecThread.Create(True);
   with fExec do begin
     FileName := sFileName;
@@ -137,13 +181,22 @@ begin
     Visible := bVisible;
     OnTerminate := TerminateEvent;
     FreeOnTerminate := True;
+    RedirectInput := bRedirectInput;
+    StartupEvent := CreateEvent(nil,False,False,nil);
     Resume;
+  end;
+  if bRedirectInput then begin
+    WaitForSingleObject(fExec.StartupEvent,INFINITE);
+    fPipe := TPipeInputThread.Create(True);
+    fPipe.WriteHandle := fExec.InputWrite;
+    fPipe.InputFile := InputFile;
+    fPipe.Execute;
   end;
 end;
 
 procedure TdevExecutor.Reset;
 begin
-  if Assigned(fExec) then
+  if Assigned(fExec) and fIsRunning then
     TerminateProcess(fExec.Process, 0);
   fIsRunning := False;
 end;
@@ -155,5 +208,58 @@ begin
     fOnTermEvent(Self);
 end;
 
-end.
+{ TPipeInputThread }
 
+procedure TPipeInputThread.WriteEndLine;
+var
+  buffer: array[0..10] of ansichar;
+  bytesWritten : cardinal;
+begin
+  buffer[0]:=#32;
+//  buffer[1]:=#0;
+  WriteFile(WriteHandle,buffer,1,bytesWritten,nil);
+end;
+procedure TPipeInputThread.PipeInput;
+const
+  BufSize = 8196;
+var
+  buffer: pAnsichar;
+  FileHandle : THandle;
+  bytesRead: cardinal;
+  bytesWritten: cardinal;
+begin
+  FileHandle := CreateFile(pAnsiChar(InputFile),GENERIC_READ, FILE_SHARE_READ,
+    nil, OPEN_EXISTING,FILE_ATTRIBUTE_NORMAL,0);
+
+  if FileHandle = INVALID_HANDLE_VALUE then begin
+    MessageDlg('Open InputFile Failed:'+SysErrorMessage(GetLastError), mtError, [mbOK], 0);
+    Exit;
+  end;
+  GetMem(buffer,BufSize+10);
+  try
+    while True do begin
+      if not ReadFile(FileHandle,buffer^,BufSize,bytesRead,nil) then begin
+        MessageDlg('Read InputFile Failed:'+SysErrorMessage(GetLastError), mtError, [mbOK], 0);
+        Exit;
+      end;
+      if bytesRead = 0 then begin
+        WriteEndLine; // write a '\n' in case of scanf waiting for it;
+        Exit;
+      end;
+      if not WriteFile(WriteHandle,buffer^,bytesRead,bytesWritten,nil) then begin
+        Exit;
+      end;
+    end;
+  finally
+    FreeMem(buffer);
+    CloseHandle(FileHandle);
+    CloseHandle(WriteHandle);
+  end;
+end;
+procedure TPipeInputThread.Execute;
+begin
+  inherited;
+  PipeInput;
+end;
+
+end.
